# docs/swagger.yaml
basePath: /
definitions:
  models.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        description: ❗ tự khai báo thay vì gorm.DeletedAt
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Laptop
        type: string
      updated_at:
        type: string
    type: object
...
  models.Product:
    properties:
      category:
        allOf:
        - $ref: '#/definitions/models.Category'
        description: sẽ dùng struct sửa như trên
      category_id:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        example: 1
        type: integer
      image:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  models.Review:
    properties:
      comment:
        example: Sản phẩm rất tốt
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 1
        type: integer
      product_id:
        example: 101
        type: integer
      rating:
        example: 4
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        example: 5
        type: integer
    type: object
  models.User:
    properties:
      address:
        type: string
      avatar:
        type: string
      bio:
        type: string
      city:
        type: string
      created_at:
        type: string
      district:
        type: string
      dob:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      job:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
      ward:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: you@example.com
    name: Developer
  description: API cho hệ thống bán hàng máy tính
  license:
    name: MIT
  termsOfService: http://localhost:8080/
  title: Computer Store API
  version: "1.0"
paths:
...
  /products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Danh sách sản phẩm
      tags:
      - Products
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Tên sản phẩm
        in: formData
        name: name
        required: true
        type: string
      - description: Mô tả
        in: formData
        name: description
        required: true
        type: string
      - description: Giá
        in: formData
        name: price
        required: true
        type: number
      - description: Số lượng
        in: formData
        name: stock
        required: true
        type: integer
      - description: ID danh mục
        in: formData
        name: category_id
        required: true
        type: integer
      - description: Ảnh sản phẩm
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Tạo sản phẩm
      tags:
      - Products
  /products/{id}:
    delete:
      parameters:
      - description: ID sản phẩm
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Xoá sản phẩm
      tags:
      - Products
    get:
      parameters:
      - description: ID sản phẩm
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Xem chi tiết sản phẩm
      tags:
      - Products
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: ID sản phẩm
        in: path
        name: id
        required: true
        type: string
      - description: Tên sản phẩm
        in: formData
        name: name
        type: string
      - description: Mô tả
        in: formData
        name: description
        type: string
      - description: Giá
        in: formData
        name: price
        type: number
      - description: Số lượng
        in: formData
        name: stock
        type: integer
      - description: ID danh mục
        in: formData
        name: category_id
        type: integer
      - description: Ảnh mới (nếu có)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cập nhật sản phẩm
      tags:
      - Products
  /profile:
    delete:
      description: Xoá tài khoản hiện tại và toàn bộ dữ liệu liên quan
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Xoá tài khoản người dùng
      tags:
      - Profile
    get:
      description: Lấy thông tin của người dùng đang đăng nhập
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lấy thông tin tài khoản
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Cập nhật thông tin tài khoản người dùng (không bao gồm avatar)
      parameters:
      - description: Thông tin người dùng
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật thông tin tài khoản
      tags:
      - Profile
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Người dùng tạo đánh giá cho một sản phẩm
      parameters:
      - description: Thông tin đánh giá
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Tạo đánh giá mới
      tags:
      - Reviews
  /reviews/{id}:
    delete:
      description: Xoá một đánh giá theo ID
      parameters:
      - description: ID đánh giá
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Xoá đánh giá
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      description: Cập nhật nội dung hoặc điểm đánh giá theo ID
      parameters:
      - description: ID đánh giá
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin đánh giá cập nhật
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Review'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật đánh giá
      tags:
      - Reviews
  /reviews/product/{product_id}:
    get:
      description: Trả về danh sách đánh giá của một sản phẩm cụ thể (bao gồm thông
        tin người dùng)
      parameters:
      - description: ID sản phẩm
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lấy đánh giá theo sản phẩm
      tags:
      - Reviews
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"